add mapper to CandidateEntityMpper to finish Find a Candidate site, starting witch User repository

+++ src/main/java/pl/devfinder/infrastructure/database/repository/jpa/CandidateJpaRepository.java
@@ -11,7 +11,7 @@ import pl.devfinder.infrastructure.database.entity.CandidateEntity;
 import java.util.List;
 
 @Repository
-public interface CandidateJpaRepository extends JpaRepository<CandidateEntity, Integer> {
+public interface CandidateJpaRepository extends JpaRepository<CandidateEntity, Long> {
 @Query("""
         SELECT candidate FROM CandidateEntity candidate WHERE candidate.status LIKE :state
         """
